+++ tests/api_resources/test_browsers.py
@@ -48,7 +48,18 @@ class TestBrowsers:
             kiosk_mode=True,
             memory="8GiB",
             name="checkout-flow-1",
-            network={"private_hosts": ["*.example.ts.net", "100.64.0.0/10"]},
+            network={
+                "private_hosts": ["*.example.ts.net", "100.64.0.0/10"],
+                "proxy_routes": [
+                    {
+                        "hosts": ["string"],
+                        "proxy": {
+                            "id": "x",
+                            "name": "x",
+                        },
+                    }
+                ],
+            },
             profile={
                 "id": "id",
                 "name": "name",
@@ -579,7 +590,18 @@ class TestAsyncBrowsers:
             kiosk_mode=True,
             memory="8GiB",
             name="checkout-flow-1",
-            network={"private_hosts": ["*.example.ts.net", "100.64.0.0/10"]},
+            network={
+                "private_hosts": ["*.example.ts.net", "100.64.0.0/10"],
+                "proxy_routes": [
+                    {
+                        "hosts": ["string"],
+                        "proxy": {
+                            "id": "x",
+                            "name": "x",
+                        },
+                    }
+                ],
+            },
             profile={
                 "id": "id",
                 "name": "name",